added isSolved parameter in struct tile_st
the "isSolved" parameter is obsolete for every single tile of the grid except for the one with [0][0] coordinates. it is meant to prevent the user from producing unintended results with the solver once the grid has already been solved (e.g. deleting every value the solver had previously found in the grid, one-by-one)